Move files to the client package
