+++ fabric-data-generation-api-v1/src/client/java/net/fabricmc/fabric/api/client/datagen/v1/provider/FabricSoundsProvider.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package net.fabricmc.fabric.api.datagen.v1.provider;
+package net.fabricmc.fabric.api.client.datagen.v1.provider;
 
 import java.nio.file.Path;
 import java.util.LinkedHashMap;
@@ -32,9 +32,8 @@ import net.minecraft.registry.RegistryWrapper;
 import net.minecraft.registry.entry.RegistryEntry;
 import net.minecraft.sound.SoundEvent;
 import net.minecraft.util.Identifier;
-
+import net.fabricmc.fabric.api.client.datagen.v1.builder.SoundTypeBuilder;
 import net.fabricmc.fabric.api.datagen.v1.FabricDataGenerator;
-import net.fabricmc.fabric.api.datagen.v1.builder.SoundTypeBuilder;
 import net.fabricmc.fabric.impl.datagen.SoundTypeBuilderImpl;
 
 /**
